+++ index.web.ts
@@ -1,3 +1,6 @@
+import { _run } from "./src/run";
+export { ClingoResult } from "./src/run";
+
 import emscripten from './clingo.js';
 const clingoModule = require('./clingo.wasm');
 
@@ -5,7 +8,7 @@ const clingoModule = require('./clingo.wasm');
 // `.wasm` file, we have to provide a `locateFile()` hook to redirect
 // to the appropriate URL.
 // More details: https://kripken.github.io/emscripten-site/docs/api_reference/module.html
-const wasmModule = emscripten({
+const webClingoModule = emscripten({
   locateFile(path) {
     if(path.endsWith('.wasm')) {
       return clingoModule;
@@ -14,7 +17,4 @@ const wasmModule = emscripten({
   }
 });
 
-
-wasmModule.onRuntimeInitialized = () => {
-  console.log(wasmModule);
-};
+export const run = _run(webClingoModule);
