+++ package.json
@@ -19,6 +19,8 @@
     "test": "ts-mocha ./test/**/*.ts"
   },
   "dependencies": {
+  },
+  "devDependencies": {
     "@types/chai": "^4.2.3",
     "@types/chai-as-promised": "^7.1.2",
     "@types/emscripten": "^1.38.0",
@@ -26,16 +28,15 @@
     "@types/node": "^12.7.8",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "mocha": "^6.2.1",
-    "ts-mocha": "^6.0.0"
-  },
-  "devDependencies": {
     "exports-loader": "^0.7.0",
     "file-loader": "^4.2.0",
     "gh-pages": "^2.1.1",
     "http-server": "^0.11.1",
+    "mocha": "^6.2.1",
     "ts-loader": "^6.1.2",
     "typescript": "^3.6.3",
+    "ts-mocha": "^6.0.0",
+    "watch": "^1.0.2",
     "webpack": "^4.41.0",
     "webpack-cli": "^3.3.9"
   }
